Move test files from 'com.lightbend.grpc' to 'akka.grpc'
For consistency
