+++ interop-tests/src/main/scala/akka/grpc/interop/GrpcClient.scala
@@ -1,0 +1,12 @@
+/*
+ * Copyright (C) 2018-2019 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.grpc.interop
+
+/**
+ * Glue code to start a gRPC client (either akka-grpc or io.grpc) to test with
+ */
+abstract class GrpcClient {
+  def run(args: Array[String]): Unit
+}
