--- interop-tests/src/main/scala/com/lightbend/grpc/interop/GrpcClient.scala
@@ -1,9 +1,0 @@
-/*
- * Copyright (C) 2018-2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package com.lightbend.grpc.interop
-
-abstract class GrpcClient {
-  def run(args: Array[String]): Unit
-}
