--- interop-tests/src/main/scala/com/lightbend/grpc/interop/IoGrpcClient.scala
@@ -1,26 +1,0 @@
-/*
- * Copyright (C) 2018-2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package com.lightbend.grpc.interop
-
-import io.grpc.internal.testing.TestUtils
-import io.grpc.testing.integration2.{ GrpcJavaClientTester, Settings, TestServiceClient }
-
-object IoGrpcClient extends GrpcClient {
-
-  override def run(args: Array[String]): Unit = {
-    TestUtils.installConscryptIfAvailable()
-    val settings = Settings.parseArgs(args)
-
-    val client = new TestServiceClient(new GrpcJavaClientTester(settings))
-    client.setUp()
-
-    try
-      client.run(settings)
-    finally {
-      client.tearDown()
-    }
-  }
-
-}
