--- interop-tests/src/test/scala/com/lightbend/grpc/interop/TestServiceImpl.scala
@@ -1,77 +1,0 @@
-/*
- * Copyright (C) 2018-2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package com.lightbend.grpc.interop
-
-import akka.NotUsed
-import akka.grpc.GrpcServiceException
-import akka.stream.Materializer
-import akka.stream.scaladsl.{ Flow, Source }
-import com.google.protobuf.ByteString
-import io.grpc.Status
-import io.grpc.testing.integration.empty.Empty
-
-import scala.collection.immutable
-import scala.concurrent.{ ExecutionContext, Future }
-
-// Generated by our plugin
-import io.grpc.testing.integration.messages._
-import io.grpc.testing.integration.test.TestService
-
-object TestServiceImpl {
-  val parametersToResponseFlow: Flow[ResponseParameters, StreamingOutputCallResponse, NotUsed] =
-    Flow[ResponseParameters]
-      .map { parameters =>
-        StreamingOutputCallResponse(
-          Some(Payload(body = ByteString.copyFrom(new Array[Byte](parameters.size)))))
-      }
-}
-
-// Implementation of the generated interface
-class TestServiceImpl(implicit ec: ExecutionContext, mat: Materializer) extends TestService {
-  import TestServiceImpl._
-
-  override def emptyCall(req: Empty) = Future.successful(Empty())
-
-  override def unaryCall(req: SimpleRequest): Future[SimpleResponse] = {
-    req.responseStatus match {
-      case None =>
-        Future.successful(
-          SimpleResponse(
-            Some(Payload(req.responseType, ByteString.copyFrom(new Array[Byte](req.responseSize))))))
-      case Some(requestStatus) =>
-        val responseStatus = Status.fromCodeValue(requestStatus.code).withDescription(requestStatus.message)
-        //  - Either one of the following works
-        // Future.failed(new GrpcServiceException(responseStatus))
-        throw new GrpcServiceException(responseStatus)
-    }
-  }
-
-  override def cacheableUnaryCall(in: SimpleRequest): Future[SimpleResponse] = ???
-
-  override def fullDuplexCall(in: Source[StreamingOutputCallRequest, NotUsed]): Source[StreamingOutputCallResponse, NotUsed] =
-    in.map(req => {
-      req.responseStatus.foreach(reqStatus =>
-        throw new GrpcServiceException(
-          Status.fromCodeValue(reqStatus.code).withDescription(reqStatus.message)))
-      req
-    }).mapConcat(
-      _.responseParameters.to[immutable.Seq]).via(parametersToResponseFlow)
-
-  override def halfDuplexCall(in: Source[StreamingOutputCallRequest, NotUsed]): Source[StreamingOutputCallResponse, NotUsed] = ???
-
-  override def streamingInputCall(in: Source[StreamingInputCallRequest, NotUsed]): Future[StreamingInputCallResponse] = {
-    in
-      .map(_.payload.map(_.body.size).getOrElse(0))
-      .runFold(0)(_ + _)
-      .map { sum ⇒
-        StreamingInputCallResponse(sum)
-      }
-  }
-
-  override def streamingOutputCall(in: StreamingOutputCallRequest): Source[StreamingOutputCallResponse, NotUsed] =
-    Source(in.responseParameters.to[immutable.Seq]).via(parametersToResponseFlow)
-
-  override def unimplementedCall(in: Empty): Future[Empty] = ???
-}
