--- interop-tests/src/test/scala/com/lightbend/grpc/interop/app/AkkaHttpServerAppScala.scala
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2018-2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package com.lightbend.grpc.interop.app
-
-import com.lightbend.grpc.interop.AkkaHttpServerProviderScala
-
-object AkkaHttpServerAppScala extends App {
-  AkkaHttpServerProviderScala.server.start()
-}
