+++ sbt-plugin/src/sbt-test/gen-scala-server/00-interop/src/test/scala/akka/grpc/GrpcInteropSpec.scala
@@ -1,0 +1,82 @@
+package akka.grpc.interop
+
+import akka.http.scaladsl.model._
+import akka.http.scaladsl.server.RouteResult.Complete
+import akka.http.scaladsl.server.{ Directive0, Directives, Route }
+import io.grpc.testing.integration.TestServiceHandlerFactory
+import io.grpc.testing.integration.test.TestService
+import org.scalatest.WordSpec
+
+import scala.collection.immutable
+
+// Generated by our plugin
+import io.grpc.testing.integration.test.TestServiceHandler
+
+class GrpcInteropSpec extends WordSpec with GrpcInteropTests with Directives {
+
+  grpcTests(IoGrpcJavaServerProvider, AkkaHttpClientProvider)
+
+  grpcTests(AkkaHttpServerProviderScala, IoGrpcJavaClientProvider)
+  grpcTests(AkkaHttpServerProviderScala, AkkaHttpClientProvider)
+
+  grpcTests(AkkaHttpServerProviderJava, IoGrpcJavaClientProvider)
+  grpcTests(AkkaHttpServerProviderJava, AkkaHttpClientProvider)
+
+  object AkkaHttpServerProviderScala extends AkkaHttpServerProvider {
+    val label: String = "akka-grpc server scala"
+    val pendingCases = Set()
+
+    val server = AkkaGrpcServerScala(implicit mat => implicit sys => {
+      implicit val ec = mat.executionContext
+
+      val requestHandler = TestServiceHandler(new TestServiceImpl())
+
+      val route: Route = (pathPrefix(TestService.name) & echoHeaders) { ctx ⇒
+        requestHandler(ctx.request).map(Complete)
+      }
+
+      Route.asyncHandler(Route.seal(route))
+    })
+
+    val echoHeaders: Directive0 = extractRequest.flatMap(request ⇒ {
+      val initialHeaderToEcho = request.headers.find(_.name() == "x-grpc-test-echo-initial")
+      val trailingHeaderToEcho = request.headers.find(_.name() == "x-grpc-test-echo-trailing-bin")
+
+      mapResponseHeaders(h ⇒ h ++ initialHeaderToEcho) & mapTrailingResponseHeaders(h ⇒ h ++ trailingHeaderToEcho)
+    })
+
+    // TODO to be moved to the runtime lib (or even akka-http itself?)
+    def mapTrailingResponseHeaders(f: immutable.Seq[HttpHeader] ⇒ immutable.Seq[HttpHeader]) =
+      mapResponse(response ⇒
+        response.withEntity(response.entity match {
+          case HttpEntity.Chunked(contentType, data) ⇒ {
+            HttpEntity.Chunked(contentType, data.map {
+              case chunk: HttpEntity.Chunk ⇒ chunk
+              case last: HttpEntity.LastChunk ⇒ HttpEntity.LastChunk(last.extension, f(last.trailer))
+            })
+          }
+          case _ ⇒
+            throw new IllegalArgumentException("Trailing response headers are only supported on Chunked responses")
+        }))
+  }
+
+  object AkkaHttpServerProviderJava extends AkkaHttpServerProvider {
+    val label: String = "akka-grpc server java"
+
+    val pendingCases =
+      Set(
+        "custom_metadata"
+      )
+
+    val server = new AkkaGrpcServerJava((mat, sys) ⇒ {
+      TestServiceHandlerFactory.create(new JavaTestServiceImpl(mat), mat, sys)
+    })
+  }
+
+  object AkkaHttpClientProvider extends AkkaHttpClientProvider {
+    val label: String = "akka-grpc scala client tester"
+
+    def client = AkkaGrpcClientScala(settings => implicit mat => implicit ec => new AkkaGrpcClientTester(settings))
+  }
+
+}
